unify delphes and cms training
