--- delphes/parameters/delphes-transformer-skipconn.yaml
@@ -1,23 +1,0 @@
-tensorflow:
-  eager: no
-
-setup:
-  train: yes
-  weights:
-  lr: 1e-6
-  batch_size: 5
-  num_events_train: 40000
-  num_events_test: 10000
-  num_epochs: 300
-  dtype: float32
-  sample_weights: none
-
-parameters:
-  model: transformer
-  num_layers: 4
-  d_model: 128
-  num_heads: 4
-  dff: 128
-  support: 32
-  skip_connection: yes
-  dropout: 0.2
