--- delphes/parameters/test.yaml
@@ -1,30 +1,0 @@
-tensorflow:
-  eager: no
-
-setup:
-  weights:
-  lr: 1e-6
-  batch_size: 5
-  num_events_train: 100
-  num_events_test: 100
-  num_epochs: 10
-  dtype: float32
-  sample_weights: none
- 
-parameters:
-  model: gnn
-  bin_size: 128
-  num_convs_id: 2
-  num_convs_reg: 2
-  num_hidden_id_enc: 2
-  num_hidden_id_dec: 2
-  num_hidden_reg_enc: 2
-  num_hidden_reg_dec: 2
-  num_neighbors: 16 
-  hidden_dim_id: 256
-  hidden_dim_reg: 256
-  distance_dim: 256
-  dropout: 0.2
-  dist_mult: 1.0
-  activation: elu
-  skip_connection: True
