+++ mlpf/tfmodel/model.py
@@ -1,53 +1,14 @@
-import os
-import sys
-import random
-os.environ["KERAS_BACKEND"] = "tensorflow"
+# This file contains the generic MLPF model definitions
+# PFNet: the GNN-based model with graph building based on LSH+kNN
+# Transformer: the transformer-based model using fast attention
+# DummyNet: simple elementwise feed forward network for cross-checking
 
-import glob
-try:
-    if not ("CUDA_VISIBLE_DEVICES" in os.environ):
-        print("importing setGPU")
-        import setGPU
-except:
-    print("Could not import setGPU, please make sure you configure CUDA_VISIBLE_DEVICES manually")
-    pass
-
-from comet_ml import Experiment
-
-import pickle
-import matplotlib.pyplot as plt
-import numpy as np
-from sklearn.metrics import confusion_matrix, accuracy_score
-import pandas
-import time
-import itertools
-import io
 import tensorflow as tf
 
-from fast_attention import Attention, SelfAttention
+from tfmodel.fast_attention import Attention, SelfAttention
 
-import sys
-#sys.path += ["/home/joosep/performer"]
-#import performer
-#import performer.networks
-#from performer.networks.linear_attention import Performer
-
-#physical_devices = tf.config.list_physical_devices('GPU')
-#tf.config.experimental.set_memory_growth(physical_devices[0], True)
-
+import numpy as np
 from numpy.lib.recfunctions import append_fields
-
-elem_labels = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11]
-class_labels = [0, 1, 2, 11, 13, 22, 130, 211]
-
-num_max_elems = 5000
-
-mult_classification_loss = 1e3
-mult_charge_loss = 1.0
-mult_energy_loss = 10.0
-mult_phi_loss = 10.0
-mult_eta_loss = 10.0
-mult_total_loss = 1e3
 
 def split_indices_to_bins(cmul, nbins, bin_size):
     bin_idx = tf.argmax(cmul, axis=-1)
@@ -89,108 +50,6 @@ def sparse_dense_matmult_batch(sp_a, b, nelem, ncol):
     elems = (tf.range(0, num_batches, delta=1, dtype=tf.int64), b)
     ret = tf.map_fn(map_function, elems, fn_output_signature=tf.TensorSpec((nelem, ncol), b.dtype), back_prop=True)
     return tf.cast(ret, dtype) 
-
-
-def summarize_dataset(dataset):
-    yclasses = []
-    nev = 0.0
-    ntot = 0.0
-    sizes = []
-
-    for X, y, w in dataset:
-        yclasses += [y[:, 0]]
-        nev += 1
-        ntot += len(y)
-        sizes += [len(y)]
-    
-    yclasses = np.concatenate(yclasses)
-    values, counts= np.unique(yclasses, return_counts=True)
-    print("nev={}".format(nev))
-    print("sizes={}".format(np.percentile(sizes, [25, 50, 95, 99])))
-    for v, c in zip(values, counts):
-        print("label={} count={} frac={:.6f}".format(class_labels[int(v)], c, c/ntot))
-
-#https://arxiv.org/pdf/1901.05555.pdf
-beta = 0.9999 #beta -> 1 means weight by inverse frequency, beta -> 0 means no reweighting
-def compute_weights_classbalanced(X, y, w):
-    wn = (1.0 - beta)/(1.0 - tf.pow(beta, w))
-    wn /= tf.reduce_sum(wn)
-    return X, y, wn
-
-#uniform weights
-def compute_weights_uniform(X, y, w):
-    wn = tf.ones_like(w)
-    wn /= tf.reduce_sum(wn)
-    return X, y, wn
-
-#weight proportional to 1/sqrt(N)
-def compute_weights_inverse(X, y, w):
-    wn = 1.0/tf.sqrt(w)
-    wn /= tf.reduce_sum(wn)
-    return X, y, wn
-
-weight_schemes = {
-    "uniform": compute_weights_uniform,
-    "inverse": compute_weights_inverse,
-    "classbalanced": compute_weights_classbalanced,
-}
-
-def load_one_file(fn):
-    Xs = []
-    ys = []
-    ys_cand = []
-    dms = []
-
-    data = pickle.load(open(fn, "rb"), encoding='iso-8859-1')
-    for event in data:
-        Xelem = event["Xelem"]
-        ygen = event["ygen"]
-        ycand = event["ycand"]
-
-        #remove PS from inputs, they don't seem to be very useful
-        msk_ps = (Xelem["typ"] == 2) | (Xelem["typ"] == 3)
-
-        Xelem = Xelem[~msk_ps]
-        ygen = ygen[~msk_ps]
-        ycand = ycand[~msk_ps]
-
-        Xelem = append_fields(Xelem, "typ_idx", np.array([elem_labels.index(int(i)) for i in Xelem["typ"]], dtype=np.float32))
-        ygen = append_fields(ygen, "typ_idx", np.array([class_labels.index(abs(int(i))) for i in ygen["typ"]], dtype=np.float32))
-        ycand = append_fields(ycand, "typ_idx", np.array([class_labels.index(abs(int(i))) for i in ycand["typ"]], dtype=np.float32))
-    
-        Xelem_flat = np.stack([Xelem[k].view(np.float32).data for k in [
-            'typ_idx',
-            'pt', 'eta', 'phi', 'e',
-            'layer', 'depth', 'charge', 'trajpoint',
-            'eta_ecal', 'phi_ecal', 'eta_hcal', 'phi_hcal',
-            'muon_dt_hits', 'muon_csc_hits']], axis=-1
-        )
-        ygen_flat = np.stack([ygen[k].view(np.float32).data for k in [
-            'typ_idx',
-            'eta', 'phi', 'e', 'charge',
-            ]], axis=-1
-        )
-        ycand_flat = np.stack([ycand[k].view(np.float32).data for k in [
-            'typ_idx',
-            'eta', 'phi', 'e', 'charge',
-            ]], axis=-1
-        )
-
-        #take care of outliers
-        Xelem_flat[np.isnan(Xelem_flat)] = 0
-        Xelem_flat[np.abs(Xelem_flat) > 1e4] = 0
-        ygen_flat[np.isnan(ygen_flat)] = 0
-        ygen_flat[np.abs(ygen_flat) > 1e4] = 0
-        ycand_flat[np.isnan(ycand_flat)] = 0
-        ycand_flat[np.abs(ycand_flat) > 1e4] = 0
-
-        Xs += [Xelem_flat[:num_max_elems]]
-        ys += [ygen_flat[:num_max_elems]]
-        ys_cand += [ycand_flat[:num_max_elems]]
-    
-    print("created {} blocks, max size {}".format(len(Xs), max([len(X) for X in Xs])))
-    return Xs, ys, ys_cand
-
 
 class InputEncoding(tf.keras.layers.Layer):
     def __init__(self, num_input_classes):
@@ -364,7 +223,6 @@ class SparseHashedNNDistance(tf.keras.layers.Layer):
         return inds, vals
 
     def construct_sparse_dm_batch(self, points):
-
         #points: (n_points, n_features) input elements for graph construction
         n_points = tf.shape(points)[0]
         n_features = tf.shape(points)[1]
@@ -451,8 +309,8 @@ def point_wise_feed_forward_network(d_model, dff, num_layers=1, activation='elu'
 #Simple message passing based on a matrix multiplication
 class PFNet(tf.keras.Model):
     def __init__(self,
-        num_input_classes=len(elem_labels),
-        num_output_classes=len(class_labels),
+        num_input_classes=8,
+        num_output_classes=3,
         num_momentum_outputs=3,
         activation=tf.nn.selu,
         hidden_dim_id=256,
@@ -526,9 +384,9 @@ class PFNet(tf.keras.Model):
         self.gnn_reg = EncoderDecoderGNN(encoding_reg, decoding_reg, dropout, activation, convs_reg, name="gnn_reg")
         self.layer_momentum = point_wise_feed_forward_network(num_momentum_outputs, hidden_dim_reg, num_layers=2, activation=activation)
 
-    def create_model(self, num_max_elems, training=True):
-        inputs = tf.keras.Input(shape=(num_max_elems,15,))
-        return tf.keras.Model(inputs=[inputs], outputs=self.call(inputs, training), name="MLPFNet")
+    # def create_model(self, num_max_elems, num_input_features, training=True):
+    #     inputs = tf.keras.Input(shape=(num_max_elems, num_input_features,))
+    #     return tf.keras.Model(inputs=[inputs], outputs=self.call(inputs, training), name="MLPFNet")
 
     def call(self, inputs, training=True):
         X = inputs
@@ -588,6 +446,8 @@ class PFNet(tf.keras.Model):
             layer.trainable = False
         self.gnn_reg.trainable = True
         self.layer_momentum.trainable = True
+
+
 
 #Transformer code from the TF example
 class EncoderLayer(tf.keras.layers.Layer):
@@ -690,13 +550,15 @@ class Decoder(tf.keras.layers.Layer):
         # x.shape == (batch_size, target_seq_len, d_model)
         return x
 
+
+
 class Transformer(tf.keras.Model):
     def __init__(self,
                 num_layers, d_model, num_heads, dff,
                 dropout=0.1,
                 support=32,
-                num_input_classes=len(elem_labels),
-                num_output_classes=len(class_labels),
+                num_input_classes=8,
+                num_output_classes=3,
                 num_momentum_outputs=3,
                 dtype=tf.dtypes.float32,
                 skip_connection=False):
@@ -756,10 +618,12 @@ class Transformer(tf.keras.Model):
         ret = tf.concat([out_id_logits, out_charge, pred_momentum], axis=-1)
         return ret
 
+
+
 class DummyNet(tf.keras.Model):
     def __init__(self,
-                num_input_classes=len(elem_labels),
-                num_output_classes=len(class_labels),
+                num_input_classes=8,
+                num_output_classes=3,
                 num_momentum_outputs=3):
         super(DummyNet, self).__init__()
 
@@ -786,422 +650,3 @@ class DummyNet(tf.keras.Model):
         ret = tf.concat([out_id_logits, out_charge, pred_momentum], axis=-1)
 
         return ret
-
-#@tf.function
-def separate_prediction(y_pred):
-    N = len(class_labels)
-    pred_id_logits = y_pred[:, :, :N]
-    pred_momentum = y_pred[:, :, N:N+3]
-    pred_charge = y_pred[:, :, N+3:N+4]
-    return pred_id_logits, pred_charge, pred_momentum
-
-#@tf.function
-def separate_truth(y_true):
-    true_id = tf.cast(y_true[:, :, :1], tf.int32)
-    true_momentum = y_true[:, :, 1:4]
-    true_charge = y_true[:, :, 4:5]
-    return true_id, true_charge, true_momentum
-
-def mse_unreduced(true, pred):
-    return tf.math.pow(true-pred,2)
-
-def msle_unreduced(true, pred):
-    return tf.math.pow(tf.math.log(tf.math.abs(true) + 1.0) - tf.math.log(tf.math.abs(pred) + 1.0), 2)
-
-def my_loss_cls(y_true, y_pred):
-    pred_id_logits, pred_charge, _ = separate_prediction(y_pred)
-    true_id, true_charge, _ = separate_truth(y_true)
-
-    true_id_onehot = tf.one_hot(tf.cast(true_id, tf.int32), depth=len(class_labels))
-    #predict the particle class labels
-    l1 = mult_classification_loss*tf.nn.softmax_cross_entropy_with_logits(true_id_onehot, pred_id_logits)
-    l3 = mult_charge_loss*mse_unreduced(true_charge, pred_charge)[:, :, 0]
-
-    loss = l1 + l3
-    return mult_total_loss*loss
-
-def my_loss_reg(y_true, y_pred):
-    _, _, pred_momentum = separate_prediction(y_pred)
-    _, true_charge, true_momentum = separate_truth(y_true)
-
-    l2_0 = mult_eta_loss*mse_unreduced(true_momentum[:, :, 0], pred_momentum[:, :, 0])
-    l2_1 = mult_phi_loss*mse_unreduced(tf.math.floormod(true_momentum[:, :, 1] - pred_momentum[:, :, 1] + np.pi, 2*np.pi) - np.pi, 0.0)
-    l2_2 = mult_energy_loss*mse_unreduced(true_momentum[:, :, 2], pred_momentum[:, :, 2])
-
-    loss = (l2_0 + l2_1 + l2_2)
-    
-    return 1e3*loss
-
-def my_loss_full(y_true, y_pred):
-    pred_id_logits, pred_charge, pred_momentum = separate_prediction(y_pred)
-    pred_id = tf.cast(tf.argmax(pred_id_logits, axis=-1), tf.int32)
-    true_id, true_charge, true_momentum = separate_truth(y_true)
-    true_id_onehot = tf.one_hot(tf.cast(true_id, tf.int32), depth=len(class_labels))
-    
-    l1 = mult_classification_loss*tf.nn.softmax_cross_entropy_with_logits(true_id_onehot, pred_id_logits)
-  
-    l2_0 = mult_eta_loss*mse_unreduced(true_momentum[:, :, 0], pred_momentum[:, :, 0])
-    l2_1 = mult_phi_loss*mse_unreduced(tf.math.floormod(true_momentum[:, :, 1] - pred_momentum[:, :, 1] + np.pi, 2*np.pi) - np.pi, 0.0)
-    l2_2 = mult_energy_loss*mse_unreduced(true_momentum[:, :, 2], pred_momentum[:, :, 2])
-
-    l2 = (l2_0 + l2_1 + l2_2)
-
-    l3 = mult_charge_loss*mse_unreduced(true_charge, pred_charge)[:, :, 0]
-    loss = l1 + l2 + l3
-
-    return mult_total_loss*loss
-
-#TODO: put these in a class
-def cls_130(y_true, y_pred):
-    pred_id_onehot, pred_charge, pred_momentum = separate_prediction(y_pred)
-    pred_id = tf.cast(tf.argmax(pred_id_onehot, axis=-1), tf.int32)
-    true_id, true_charge, true_momentum = separate_truth(y_true)
-
-    msk_true = true_id[:, :, 0] == class_labels.index(130)
-    msk_pos = pred_id == class_labels.index(130)
-    num_true_pos = tf.reduce_sum(tf.cast(msk_true&msk_pos, tf.float32))
-    num_true = tf.reduce_sum(tf.cast(msk_true, tf.float32))
-    return num_true_pos/num_true
-
-def cls_211(y_true, y_pred):
-    pred_id_onehot, pred_charge, pred_momentum = separate_prediction(y_pred)
-    pred_id = tf.cast(tf.argmax(pred_id_onehot, axis=-1), tf.int32)
-    true_id, true_charge, true_momentum = separate_truth(y_true)
-
-    msk_true = true_id[:, :, 0] == class_labels.index(211)
-    msk_pos = pred_id == class_labels.index(211)
-    num_true_pos = tf.reduce_sum(tf.cast(msk_true&msk_pos, tf.float32))
-    num_true = tf.reduce_sum(tf.cast(msk_true, tf.float32))
-
-    return num_true_pos/num_true
-
-def cls_22(y_true, y_pred):
-    pred_id_onehot, pred_charge, pred_momentum = separate_prediction(y_pred)
-    pred_id = tf.cast(tf.argmax(pred_id_onehot, axis=-1), tf.int32)
-    true_id, true_charge, true_momentum = separate_truth(y_true)
-
-    msk_true = true_id[:, :, 0] == class_labels.index(22)
-    msk_pos = pred_id == class_labels.index(22)
-    num_true_pos = tf.reduce_sum(tf.cast(msk_true&msk_pos, tf.float32))
-    num_true = tf.reduce_sum(tf.cast(msk_true, tf.float32))
-
-    return num_true_pos/num_true
-
-def cls_11(y_true, y_pred):
-    pred_id_onehot, pred_charge, pred_momentum = separate_prediction(y_pred)
-    pred_id = tf.cast(tf.argmax(pred_id_onehot, axis=-1), tf.int32)
-    true_id, true_charge, true_momentum = separate_truth(y_true)
-
-    msk_true = true_id[:, :, 0] == class_labels.index(11)
-    msk_pos = pred_id == class_labels.index(11)
-    num_true_pos = tf.reduce_sum(tf.cast(msk_true&msk_pos, tf.float32))
-    num_true = tf.reduce_sum(tf.cast(msk_true, tf.float32))
-
-    return num_true_pos/num_true
-
-def cls_13(y_true, y_pred):
-    pred_id_onehot, pred_charge, pred_momentum = separate_prediction(y_pred)
-    pred_id = tf.cast(tf.argmax(pred_id_onehot, axis=-1), tf.int32)
-    true_id, true_charge, true_momentum = separate_truth(y_true)
-
-    msk_true = true_id[:, :, 0] == class_labels.index(13)
-    msk_pos = pred_id == class_labels.index(13)
-    num_true_pos = tf.reduce_sum(tf.cast(msk_true&msk_pos, tf.float32))
-    num_true = tf.reduce_sum(tf.cast(msk_true, tf.float32))
-
-    return num_true_pos/num_true
-
-#@tf.function
-def num_pred(y_true, y_pred):
-    pred_id_onehot, pred_charge, pred_momentum = separate_prediction(y_pred)
-    pred_id = tf.cast(tf.argmax(pred_id_onehot, axis=-1), tf.int32)
-    true_id, true_charge, true_momentum = separate_truth(y_true)
-
-    ntrue = tf.reduce_sum(tf.cast(true_id[:, :, 0]!=0, tf.int32))
-    npred = tf.reduce_sum(tf.cast(pred_id!=0, tf.int32))
-    return tf.cast(ntrue - npred, tf.float32)
-
-def accuracy(y_true, y_pred):
-    pred_id_onehot, pred_charge, pred_momentum = separate_prediction(y_pred)
-    pred_id = tf.cast(tf.argmax(pred_id_onehot, axis=-1), tf.int32)
-    true_id, true_charge, true_momentum = separate_truth(y_true)
-
-    is_true = true_id[:, :, 0]!=0
-    is_same = true_id[:, :, 0] == pred_id
-
-    acc = tf.reduce_sum(tf.cast(is_true&is_same, tf.int32)) / tf.reduce_sum(tf.cast(is_true, tf.int32))
-    return tf.cast(acc, tf.float32)
-
-#@tf.function
-def eta_resolution(y_true, y_pred):
-    pred_id_onehot, pred_charge, pred_momentum = separate_prediction(y_pred)
-    pred_id = tf.cast(tf.argmax(pred_id_onehot, axis=-1), tf.int32)
-    true_id, true_charge, true_momentum = separate_truth(y_true)
-
-    msk = true_id[:, :, 0]!=0
-    return tf.reduce_mean(mse_unreduced(true_momentum[msk][:, 0], pred_momentum[msk][:, 0]))
-
-#@tf.function
-def phi_resolution(y_true, y_pred):
-    pred_id_onehot, pred_charge, pred_momentum = separate_prediction(y_pred)
-    pred_id = tf.cast(tf.argmax(pred_id_onehot, axis=-1), tf.int32)
-    true_id, true_charge, true_momentum = separate_truth(y_true)
-
-    msk = true_id[:, :, 0]!=0
-    return tf.reduce_mean(mse_unreduced(tf.math.floormod(true_momentum[msk][:, 1] - pred_momentum[msk][:, 1] + np.pi, 2*np.pi) - np.pi, 0.0))
-
-#@tf.function(experimental_relax_shapes=True)
-def energy_resolution(y_true, y_pred):
-    pred_id_onehot, pred_charge, pred_momentum = separate_prediction(y_pred)
-    pred_id = tf.cast(tf.argmax(pred_id_onehot, axis=-1), tf.int32)
-    true_id, true_charge, true_momentum = separate_truth(y_true)
-
-    msk = true_id[:, :, 0]!=0
-    return tf.reduce_mean(mse_unreduced(true_momentum[msk][:, 2], pred_momentum[msk][:, 2]))
-
-def get_unique_run():
-    previous_runs = os.listdir('experiments')
-    if len(previous_runs) == 0:
-        run_number = 1
-    else:
-        run_number = max([int(s.split('run_')[1]) for s in previous_runs]) + 1
-    return run_number
-
-def parse_args():
-    import argparse
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--model", type=str, default="PFNet", help="type of model to train", choices=["PFNet"])
-    parser.add_argument("--ntrain", type=int, default=100, help="number of training events")
-    parser.add_argument("--ntest", type=int, default=100, help="number of testing events")
-    parser.add_argument("--nepochs", type=int, default=100, help="number of training epochs")
-    parser.add_argument("--hidden-dim-id", type=int, default=256, help="hidden dimension")
-    parser.add_argument("--hidden-dim-reg", type=int, default=256, help="hidden dimension")
-    parser.add_argument("--batch-size", type=int, default=1, help="number of events in training batch")
-    parser.add_argument("--num-convs-id", type=int, default=1, help="number of convolution layers")
-    parser.add_argument("--num-convs-reg", type=int, default=1, help="number of convolution layers")
-    parser.add_argument("--num-hidden-id-enc", type=int, default=2, help="number of encoder layers for multiclass")
-    parser.add_argument("--num-hidden-id-dec", type=int, default=2, help="number of decoder layers for multiclass")
-    parser.add_argument("--num-hidden-reg-enc", type=int, default=2, help="number of encoder layers for regression")
-    parser.add_argument("--num-hidden-reg-dec", type=int, default=2, help="number of decoder layers for regression")
-    parser.add_argument("--num-neighbors", type=int, default=5, help="number of knn neighbors")
-    parser.add_argument("--distance-dim", type=int, default=256, help="distance dimension")
-    parser.add_argument("--bin-size", type=int, default=100, help="number of points per LSH bin")
-    parser.add_argument("--dropout", type=float, default=0.1, help="Dropout rate")
-    parser.add_argument("--dist-mult", type=float, default=1.0, help="Exponential multiplier")
-    parser.add_argument("--target", type=str, choices=["cand", "gen"], help="Regress to PFCandidates or GenParticles", default="gen")
-    parser.add_argument("--weights", type=str, choices=["uniform", "inverse", "classbalanced"], help="Sample weighting scheme to use", default="inverse")
-    parser.add_argument("--name", type=str, default=None, help="where to store the output")
-    parser.add_argument("--convlayer", type=str, default="sgconv", choices=["sgconv", "ghconv"], help="Type of graph convolutional layer")
-    parser.add_argument("--load", type=str, default=None, help="model to load")
-    parser.add_argument("--datapath", type=str, help="Input data path", required=True)
-    parser.add_argument("--lr", type=float, default=1e-5, help="learning rate")
-    parser.add_argument("--lr-decay", type=float, default=0.0, help="learning rate decay")
-    parser.add_argument("--train-cls", action="store_true", help="Train only the classification part")
-    parser.add_argument("--train-reg", action="store_true", help="Train only the regression part")
-    parser.add_argument("--cosine-dist", action="store_true", help="Use cosine distance")
-    parser.add_argument("--eager", action="store_true", help="Run in eager mode for debugging")
-    args = parser.parse_args()
-    return args
-
-def assign_label(pred_id_onehot_linear):
-    ret2 = np.argmax(pred_id_onehot_linear, axis=-1)
-    return ret2
-
-def prepare_df(model, data, outdir, target, save_raw=False):
-    print("prepare_df")
-
-    dfs = []
-    for iev, d in enumerate(data):
-        if iev%50==0:
-            tf.print(".", end="")
-        X, y, w = d
-        pred = model(X, training=False).numpy()
-        pred_id_onehot, pred_charge, pred_momentum = separate_prediction(pred)
-        pred_id = assign_label(pred_id_onehot).flatten()
- 
-        if save_raw:
-            np.savez_compressed("ev_{}.npz".format(iev), X=X.numpy(), y=y.numpy(), w=w.numpy(), y_pred=pred)
-
-        pred_charge = pred_charge[:, :, 0].flatten()
-        pred_momentum = pred_momentum.reshape((pred_momentum.shape[0]*pred_momentum.shape[1], pred_momentum.shape[2]))
-
-        true_id, true_charge, true_momentum = separate_truth(y)
-        true_id = true_id.numpy()[:, :, 0].flatten()
-        true_charge = true_charge.numpy()[:, :, 0].flatten()
-        true_momentum = true_momentum.numpy().reshape((true_momentum.shape[0]*true_momentum.shape[1], true_momentum.shape[2]))
-       
-        df = pandas.DataFrame()
-        df["pred_pid"] = np.array([int(class_labels[p]) for p in pred_id])
-        df["pred_eta"] = np.array(pred_momentum[:, 0], dtype=np.float64)
-        df["pred_phi"] = np.array(pred_momentum[:, 1], dtype=np.float64)
-        df["pred_e"] = np.array(pred_momentum[:, 2], dtype=np.float64)
-
-        df["{}_pid".format(target)] = np.array([int(class_labels[p]) for p in true_id])
-        df["{}_eta".format(target)] = np.array(true_momentum[:, 0], dtype=np.float64)
-        df["{}_phi".format(target)] = np.array(true_momentum[:, 1], dtype=np.float64)
-        df["{}_e".format(target)] = np.array(true_momentum[:, 2], dtype=np.float64)
-
-        df["iev"] = iev
-        dfs += [df]
-    df = pandas.concat(dfs, ignore_index=True)
-    fn = outdir + "/df.pkl.bz2"
-    df.to_pickle(fn)
-    print("prepare_df done", fn)
-
-def plot_to_image(figure):
-    """Converts the matplotlib plot specified by 'figure' to a PNG image and
-    returns it. The supplied figure is closed and inaccessible after this call."""
-    # Save the plot to a PNG in memory.
-    buf = io.BytesIO()
-    plt.savefig(buf, format='png')
-    # Closing the figure prevents it from being displayed directly inside
-    # the notebook.
-    buf.seek(0)
-    # Convert PNG buffer to TF image
-    image = tf.image.decode_png(buf.getvalue(), channels=4)
-    # Add the batch dimension
-    image = tf.expand_dims(image, 0)
-    return image
-
-def load_dataset_ttbar(datapath, target):
-    from tf_data import _parse_tfr_element
-    path = datapath + "/tfr/{}/*.tfrecords".format(target)
-    tfr_files = glob.glob(path)
-    if len(tfr_files) == 0:
-        raise Exception("Could not find any files in {}".format(path))
-    dataset = tf.data.TFRecordDataset(tfr_files).map(_parse_tfr_element, num_parallel_calls=tf.data.experimental.AUTOTUNE)
-    return dataset
-
-if __name__ == "__main__":
-    args = parse_args()
-    print(args)
-    
-    experiment = Experiment(project_name="particleflow_tf")
-
-    #tf.debugging.enable_check_numerics()
-    tf.config.run_functions_eagerly(args.eager)
-
-    #batch size for loading data must be configured according to the number of distributed GPUs 
-    global_batch_size = args.batch_size
-    try:
-        num_gpus = len(os.environ["CUDA_VISIBLE_DEVICES"].split(","))
-        print("num_gpus=", num_gpus)
-        if num_gpus > 1:
-            strategy = tf.distribute.MirroredStrategy()
-            global_batch_size = num_gpus * args.batch_size
-        else:
-            strategy = tf.distribute.OneDeviceStrategy("gpu:0")
-    except Exception as e:
-        print("fallback to CPU")
-        strategy = tf.distribute.OneDeviceStrategy("cpu")
-
-    filelist = sorted(glob.glob(args.datapath + "/raw/*.pkl"))[:args.ntrain+args.ntest]
-
-    dataset = load_dataset_ttbar(args.datapath, args.target)
-
-    #create padded input data
-    ps = (tf.TensorShape([num_max_elems, 15]), tf.TensorShape([num_max_elems, 5]), tf.TensorShape([num_max_elems, ]))
-    ds_train = dataset.take(args.ntrain).map(weight_schemes[args.weights]).padded_batch(global_batch_size, padded_shapes=ps)
-    ds_test = dataset.skip(args.ntrain).take(args.ntest).map(weight_schemes[args.weights]).padded_batch(global_batch_size, padded_shapes=ps)
-
-    #repeat needed for keras api
-    ds_train_r = ds_train.repeat(args.nepochs)
-    ds_test_r = ds_test.repeat(args.nepochs)
-
-    if args.lr_decay > 0:
-        lr_schedule = tf.keras.optimizers.schedules.ExponentialDecay(
-            args.lr,
-            decay_steps=10*int(args.ntrain/global_batch_size),
-            decay_rate=args.lr_decay
-        )
-    else:
-        lr_schedule = args.lr
-
-    loss_fn = my_loss_full
-
-    with strategy.scope():
-        opt = tf.keras.optimizers.Adam(learning_rate=lr_schedule)
-
-        model = PFNet(
-            hidden_dim_id=args.hidden_dim_id,
-            hidden_dim_reg=args.hidden_dim_reg,
-            num_convs_id=args.num_convs_id,
-            num_convs_reg=args.num_convs_reg,
-            num_hidden_id_enc=args.num_hidden_id_enc,
-            num_hidden_id_dec=args.num_hidden_id_dec,
-            num_hidden_reg_enc=args.num_hidden_reg_enc,
-            num_hidden_reg_dec=args.num_hidden_reg_dec,
-            distance_dim=args.distance_dim,
-            convlayer=args.convlayer,
-            dropout=args.dropout,
-            bin_size=args.bin_size,
-            num_neighbors=args.num_neighbors,
-            dist_mult=args.dist_mult
-        )
-
-        if args.train_cls:
-            loss_fn = my_loss_cls
-            model.set_trainable_classification()
-        elif args.train_reg:
-            loss_fn = my_loss_reg
-            model.set_trainable_regression()
-
-        model(np.random.randn(args.batch_size, num_max_elems, 15).astype(np.float32))
-        if not args.eager:
-            model = model.create_model(num_max_elems)
-            model.summary()
-
-    if not os.path.isdir("experiments"):
-        os.makedirs("experiments")
-
-    if args.name is None:
-        args.name =  'run_{:02}'.format(get_unique_run())
-
-    outdir = 'experiments/' + args.name
-
-    if os.path.isdir(outdir):
-        print("Output directory exists: {}".format(outdir), file=sys.stderr)
-        sys.exit(1)
-
-    print(outdir)
-    callbacks = []
-    tb = tf.keras.callbacks.TensorBoard(
-        log_dir=outdir, histogram_freq=1, write_graph=False, write_images=False,
-        update_freq='epoch',
-        #profile_batch=(10,40),
-        profile_batch=0,
-    )
-    tb.set_model(model)
-    callbacks += [tb]
-
-    terminate_cb = tf.keras.callbacks.TerminateOnNaN()
-    callbacks += [terminate_cb]
-
-    cp_callback = tf.keras.callbacks.ModelCheckpoint(
-        filepath=outdir + "/weights.{epoch:02d}-{val_loss:.6f}.hdf5",
-        save_weights_only=True,
-        verbose=0
-    )
-    cp_callback.set_model(model)
-    callbacks += [cp_callback]
-
-    with strategy.scope():
-        model.compile(optimizer=opt, loss=loss_fn,
-            metrics=[accuracy, cls_130, cls_211, cls_22, energy_resolution, eta_resolution, phi_resolution],
-            sample_weight_mode="temporal")
-
-        if args.load:
-            #ensure model input size is known
-            for X, y, w in ds_train:
-                model(X)
-                break
-   
-            model.load_weights(args.load)
-
-        if args.nepochs > 0:
-            ret = model.fit(ds_train_r,
-                validation_data=ds_test_r, epochs=args.nepochs,
-                steps_per_epoch=args.ntrain/global_batch_size, validation_steps=args.ntest/global_batch_size,
-                verbose=True,
-                callbacks=callbacks
-            )
