+++ parameters/delphes-dense.yaml
