+++ parameters/delphes-gnn.yaml
