+++ parameters/delphes-transformer-skipconn.yaml
@@ -1,0 +1,45 @@
+backend: tensorflow
+
+dataset:
+  schema: delphes
+  num_input_features: 12
+  num_output_features: 7
+  num_input_classes: 3
+  num_output_classes: 6
+  num_momentum_outputs: 5
+  padded_num_elem_size: 6400
+  momentum_loss_coefs:
+    - 1.0
+    - 1.0
+    - 1.0
+    - 1.0
+    - 1.0
+  raw_path: delphes/out/pythia8_ttbar/*.pkl.bz2
+  processed_path: delphes/out/pythia8_ttbar/tfr/*.tfrecords
+  num_files_per_chunk: 5
+  test_file_pattern: tev14_pythia8_ttbar_[8-9]_\d+\.pkl\.bz2
+
+tensorflow:
+  eager: no
+
+setup:
+  train: yes
+  weights:
+  lr: 1e-6
+  batch_size: 5
+  num_events_train: 40000
+  num_events_test: 10000
+  num_epochs: 300
+  dtype: float32
+  sample_weights: none
+  trainable: all
+
+parameters:
+  model: transformer
+  num_layers: 4
+  d_model: 128
+  num_heads: 4
+  dff: 128
+  support: 32
+  skip_connection: yes
+  dropout: 0.2
